Store to Gist on action
Also solves an burn issue when you burn the last card in a pile - i.e. removing the array.

The GET request to gists is done as the raw, public content you can get is cached with a 5 minute cache. Meaning, you could save stuff, but you cannot load that until 5 minutes after. One issue with the updates on action is the rate limiting. After playing a bit with this, I encountered 403s due to rate limiting. Seems there are about 5000 requests for each user. However, I am able to GET stuff, but not patch..